instance info into telemetry
