+++ packages/web-console/src/store/Telemetry/epics.ts
@@ -44,6 +44,7 @@ import * as QuestDB from "../../utils/questdb"
 import { getValue } from "../../utils/localStorage"
 import { StoreKey } from "../../utils/localStorage/types"
 import { AuthPayload } from "../../modules/OAuth2/types"
+import { sendEntTelemetry } from "../../utils/telemetry";
 
 const quest = new QuestDB.Client()
 
@@ -100,17 +101,8 @@ export const getRemoteConfig: Epic<StoreAction, TelemetryAction, StoreShape> = (
     withLatestFrom(state$),
     switchMap(([_, state]) => {
       const config = selectors.telemetry.getConfig(state)
-
-      const releaseType = getValue(StoreKey.RELEASE_TYPE);
-      if (releaseType === "EE" || config?.enabled) {
-        fetch(`${API}/add-ent`, {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({ ...config, releaseType }),
-        }).catch( () => {
-        })
+      if (config) {
+          sendEntTelemetry(config)
       }
 
       if (config?.enabled) {
